refactor: move components and utils outside app directory
