+++ src/app/glossary/page.mdx
@@ -1,5 +1,5 @@
 import Glossary from '@/data/glossary.md';
-import MdxLayout from '@/app/components/MdxLayout';
+import MdxLayout from '@/components/MdxLayout';
 import 'highlight.js/styles/github.css';
 
 export default function Page() {
