+++ src/app/philosophy/page.mdx
@@ -1,5 +1,5 @@
 import Philosophy from '@/data/philosophy.md';
-import MdxLayout from '@/app/components/MdxLayout';
+import MdxLayout from '@/components/MdxLayout';
 
 export default function Page() {
   return (
