+++ src/components/PersistentDetails.tsx
