+++ src/components/essay/Note.tsx
