+++ src/components/essay/Warning.tsx
