+++ src/components/essay/season04/fp1100-beta-certainty/page.tsx
@@ -1,5 +1,5 @@
-import Narration from '@/app/components/essay/Narration';
-import EssayNavigation from '@/app/components/essay/EssayNavigation';
+import Narration from '@/components/essay/Narration';
+import EssayNavigation from '@/components/essay/EssayNavigation';
 import { User } from '../../User';
 import { AI } from '../../AI';
